requirements.txt
functions-framework==3.*

# Application dependencies
google-genai
google-cloud-aiplatform
requests